First attempt at browserify and aw-client-js
