+++ .gitmodules
@@ -1,3 +1,6 @@
 [submodule "media"]
 	path = media
 	url = https://github.com/ActivityWatch/media.git
+[submodule "aw-client-js"]
+	path = aw-client-js
+	url = git@github.com:ActivityWatch/aw-client-js.git
