+++ package.json
@@ -1,0 +1,8 @@
+{
+    "dependencies": {
+        "browserify": "^16.2.2"
+    },
+    "scripts": {
+        "build": "browserify src/*.js > out/app.js"
+    }
+}
