+++ static/popup.html
@@ -4,10 +4,9 @@
   <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
   <title>ActivityWatch Popup</title>
 
-  <link href="./popup_files/style.css" rel="stylesheet" type="text/css">
+  <link href="./style.css" rel="stylesheet" type="text/css">
 
-  <script src="./client.js"></script>
-  <script src="./popup_files/popup.js"></script>
+  <script src="/out/app.js"></script>
 </head>
 
 <body>
